pyproject.toml
[tool.poetry]
name = "s3static"
version = "0.1.0"
description = ""
authors = ["gujamark <guja.markozashvili.1@btu.edu.ge>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.26.113"
jmespath = "^1.0.1"
python-dotenv = "^1.0.0"
python-magic = "^0.4.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
